+++ R/contrast.R
@@ -313,8 +313,10 @@ contrast.emmGrid = function(object, method = "eff", interaction = FALSE,
     cmat = method(levs, ...)
     if (!is.data.frame(cmat))
         stop("Contrast function must provide a data.frame")
-    else if(ncol(cmat) == 0)
+    else if(ncol(cmat) == 0) {
         cmat = data.frame(`(nothing)` = rep(NA, nrow(args)), check.names = FALSE)
+        adjust = "none"
+    }
     # warning("No contrasts were generated! Perhaps only one emmean is involved.\n",
     #      "  This can happen, for example, when your predictors are not factors.")
     else if (nrow(cmat) != nrow(args))